Gradle Update, Google Maps API v2 & Database Export
Project changed to gradle build system as I have switched to Android Studio for my main IDE,
this change is quite significant in terms of layout of the application source code and the
abilities provided by gradle.

Google Maps API v2 support added to enable use of new features and gain an API Key for the
application as API v1 keys are no longer available, signal strength data is now correctly
overlayed on the map as well. Minor tweaks are required but currently everything is functioning
correctly, the next items to add could include markers to identify tracked CellID's and user
options to modify map preferences.

Database export corrected and changed to use the OpenCV library to produce CSV copies of the database tables.