Silent Sms Detection (Class 0), Database Restore, Progress Bar
Sms Receiver will monitor for Class 0 messages and provide an alert when received,
details are tracked to a database table and displayed in a notification fragment.

Restore database tables using exported CSV files this will also be added into the
Db Adapter shortly to allow data to persist across database updates. Database retore added to the RequestTask class so it can be executed in the background.

Progress bar added to provide more feedback about background operations.